fix: make sure the vite server starts on a new port
