+++ npm/vite-dev-server/src/startServer.ts
@@ -1,20 +1,34 @@
 import Debug from 'debug'
-import { StartDevServer } from '.'
-import { createServer, ViteDevServer, InlineConfig } from 'vite'
+import { createServer, ViteDevServer, InlineConfig, UserConfig } from 'vite'
 import { dirname, resolve } from 'path'
+import getPort from 'get-port'
 import { makeCypressPlugin } from './makeCypressPlugin'
 
 const debug = Debug('cypress:vite-dev-server:start')
 
-const resolveServerConfig = ({ viteConfig, options }: StartDevServer) => {
+interface Options {
+  specs: Cypress.Cypress['spec'][]
+  config: Record<string, string>
+  devServerEvents: EventEmitter
+  [key: string]: unknown
+}
+
+export interface StartDevServer {
+  /* this is the Cypress options object */
+  options: Options
+  /* support passing a path to the user's webpack config */
+  viteConfig?: UserConfig // TODO: implement taking in the user's vite configuration. Right now we don't
+}
+
+const resolveServerConfig = async ({ viteConfig, options }: StartDevServer): Promise<InlineConfig> => {
   const { projectRoot, supportFile } = options.config
 
-  const requiredOptions = {
+  const requiredOptions: InlineConfig = {
     base: '/__cypress/src/',
     root: projectRoot,
   }
 
-  const finalConfig = { ...viteConfig, ...requiredOptions }
+  const finalConfig: InlineConfig = { ...viteConfig, ...requiredOptions }
 
   finalConfig.plugins = [...(viteConfig.plugins || []), makeCypressPlugin(projectRoot, supportFile, options.devServerEvents)]
 
@@ -26,6 +40,10 @@ const resolveServerConfig = ({ viteConfig, options }: StartDevServer) => {
     ...finalConfig.resolve.alias,
     '@vue/compiler-core': resolve(dirname(require.resolve('@vue/compiler-core')), 'dist', 'compiler-core.cjs.js'),
   },
+
+  finalConfig.server = finalConfig.server || {}
+
+  finalConfig.server.port = await getPort({ port: 3000, host: 'localhost' }),
 
   debug(`the resolved server config is ${JSON.stringify(finalConfig, null, 2)}`)
 
@@ -39,7 +57,7 @@ export async function start (devServerOptions: StartDevServer): Promise<ViteDevS
   }
 
   debug('starting vite dev server')
-  const resolvedConfig = resolveServerConfig(devServerOptions)
+  const resolvedConfig = await resolveServerConfig(devServerOptions)
 
   return createServer(resolvedConfig)
 }
